requirements.txt
django==1.4.2
django_dajaxice
